// shared/tcevinv/tcevinv-types.ts
export interface ITcevinv {
    idevento: number;
    tipoevento: string;
    tipodoc: string;
    iddoc: number;
    codsisgen: string;
    codsisdest: string;
    fecevento: string;
    stsevento: string;
    fecsts: string;
    usrcre: string;
    usrsts: string;
    ideventodev: null | number;
    codundorig: string;
    codundadmorig: string;
    codundadmpro: string;
    raw_rnum_: number;
    mensaje: string;
    descdoc: string;
  }